# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: Eatda API Documentation
  description: Eatda(잇다) 서비스의 API 명세서입니다.
  version: 0.0.1
servers:
- url: http://localhost:8080
tags: []
paths:
  /api/v1/banners:
    get:
      tags:
      - banner
      summary: 배너 전체 조회
      description: 배너 전체 조회
      operationId: banner-findAll
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                banner-findAll:
                  value: "{\"message\":null,\"body\":[{\"id\":1,\"link\":\"https://career.programmers.co.kr/competitions/3353\"\
                    ,\"imageAddress\":\"/banner/232D8241-C6A9-4AD9-B0EA-56F6DD24BADF.png\"\
                    ,\"expiredAt\":\"2023-11-21T14:11:34.639184Z\",\"createdAt\":\"\
                    2023-11-20T14:11:52.114906Z\"}]}"
    post:
      tags:
      - banner
      summary: 배너 생성
      description: 배너 생성
      operationId: banner-create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              banner-create:
                value: "{\"link\":\"https://career.programmers.co.kr/competitions/3353\"\
                  ,\"imageAddress\":\"/banner/232D8241-C6A9-4AD9-B0EA-56F6DD24BADF.png\"\
                  ,\"expiredAt\":1700575894.639184000}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                banner-create:
                  value: "{\"message\":null,\"body\":4843367494315138152}"
  /api/v1/categories:
    get:
      tags:
      - category
      summary: 카테고리 전체조회
      description: 카테고리 전체조회
      operationId: category-findAll
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                category-findAll:
                  value: "{\"message\":null,\"body\":[{\"id\":8608100422551522095,\"\
                    name\":\"87de5dc8-c3b8-419e-ae63-9e2878482a39\"},{\"id\":2343467381135403541,\"\
                    name\":\"9722e43c-bf83-4a33-817b-fec028419260\"},{\"id\":4900245127042464096,\"\
                    name\":\"32eda463-06f4-43c8-b953-05db6ae6295b\"}]}"
    post:
      tags:
      - category
      summary: 카테고리 생성
      description: 카테고리 생성
      operationId: category-create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              category-create:
                value: "{\"name\":\"validName\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                category-create:
                  value: "{\"message\":null,\"body\":9028569975075608962}"
  /api/v1/notices:
    get:
      tags:
      - notice
      summary: 공지사항 전체조회
      description: 공지사항 전체조회
      operationId: notice-findAll
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                notice-findAll:
                  value: "{\"message\":null,\"body\":[{\"id\":1,\"title\":\"공지사항 제\
                    목\",\"content\":\"공지사항 내용\",\"createdAt\":\"2023-11-20T14:11:52.925060Z\"\
                    }]}"
    post:
      tags:
      - notice
      summary: 공지사항 생성
      description: 공지사항 생성
      operationId: notice-create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              notice-create:
                value: "{\"title\":\"title\",\"content\":\"content\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                notice-create:
                  value: "{\"message\":null,\"body\":2058921043891802126}"
  /api/v1/stores:
    get:
      tags:
      - store
      summary: 커서 기반 카테고리 가게 조회
      description: 커서 기반 카테고리 가게 조회
      operationId: store-findAllByCursor
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                store-findAllByCursor:
                  value: "{\"message\":null,\"body\":{\"content\":[{\"id\":1,\"displayedName\"\
                    :\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null},{\"id\":2,\"displayedName\":\"띵지대\",\"address\"\
                    :\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"location\":null},{\"\
                    id\":3,\"displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로\
                    \ 34\",\"imageAddress\":null,\"location\":null},{\"id\":4,\"displayedName\"\
                    :\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null},{\"id\":5,\"displayedName\":\"띵지대\",\"address\"\
                    :\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"location\":null},{\"\
                    id\":6,\"displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로\
                    \ 34\",\"imageAddress\":null,\"location\":null},{\"id\":7,\"displayedName\"\
                    :\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null},{\"id\":8,\"displayedName\":\"띵지대\",\"address\"\
                    :\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"location\":null},{\"\
                    id\":9,\"displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로\
                    \ 34\",\"imageAddress\":null,\"location\":null},{\"id\":10,\"\
                    displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"\
                    imageAddress\":null,\"location\":null},{\"id\":11,\"displayedName\"\
                    :\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null},{\"id\":12,\"displayedName\":\"띵지대\",\"address\"\
                    :\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"location\":null},{\"\
                    id\":13,\"displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골\
                    로 34\",\"imageAddress\":null,\"location\":null},{\"id\":14,\"\
                    displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"\
                    imageAddress\":null,\"location\":null},{\"id\":15,\"displayedName\"\
                    :\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null},{\"id\":16,\"displayedName\":\"띵지대\",\"address\"\
                    :\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"location\":null},{\"\
                    id\":17,\"displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골\
                    로 34\",\"imageAddress\":null,\"location\":null},{\"id\":18,\"\
                    displayedName\":\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"\
                    imageAddress\":null,\"location\":null},{\"id\":19,\"displayedName\"\
                    :\"띵지대\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null},{\"id\":20,\"displayedName\":\"띵지대\",\"address\"\
                    :\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"location\":null}],\"\
                    pageable\":\"INSTANCE\",\"first\":true,\"last\":true,\"size\"\
                    :20,\"number\":0,\"sort\":{\"empty\":true,\"unsorted\":true,\"\
                    sorted\":false},\"numberOfElements\":20,\"empty\":false}}"
    post:
      tags:
      - store
      summary: 가게 생성
      description: 가게 생성
      operationId: store-create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              store-create:
                value: "{\"name\":\"명지대학교\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"\
                  displayName\":null,\"phoneNumber\":null,\"imageAddress\":null,\"\
                  location\":null}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                store-create:
                  value: "{\"message\":null,\"body\":1}"
  /api/v1/banners/{bannerId}:
    delete:
      tags:
      - banner
      summary: 배너 삭제
      description: 배너 삭제
      operationId: banner-delete
      parameters:
      - name: bannerId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - banner
      summary: 배너 수정
      description: 배너 수정
      operationId: banner-update
      parameters:
      - name: bannerId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              banner-update:
                value: "{\"link\":\"https://career.programmers.co.kr/competitions/3353\"\
                  ,\"imageAddress\":\"/banner/232D8241-C6A9-4AD9-B0EA-56F6DD24BADF.png\"\
                  ,\"expiredAt\":1700575894.639184000}"
      responses:
        "200":
          description: "200"
  /api/v1/categories/{category-id}:
    delete:
      tags:
      - category
      summary: 카테고리 삭제
      description: 카테고리 삭제
      operationId: category-delete
      parameters:
      - name: category-id
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - category
      summary: 카테고리 이름 수정
      description: 카테고리 이름 수정
      operationId: category-updateName
      parameters:
      - name: category-id
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              category-updateName:
                value: "{\"name\":\"newName\"}"
      responses:
        "200":
          description: "200"
  /api/v1/categories/{categoryId}:
    get:
      tags:
      - category
      summary: 카테고리 단건 조회
      description: 카테고리 단건 조회
      operationId: category-findById
      parameters:
      - name: categoryId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                category-findById:
                  value: "{\"message\":null,\"body\":{\"id\":4631017781137326803,\"\
                    name\":\"뉴ef8845e-4089-41a4-a393-d00aa8386d85\"}}"
  /api/v1/menu/{menuId}:
    get:
      tags:
      - menu
      summary: 메뉴 조회
      description: 메뉴 조회
      operationId: menu-findById뉴
      parameters:
      - name: menuId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                menu-findById:
                  value: "{\"message\":null,\"body\":{\"id\":1,\"name\":\"menu\",\"\
                    price\":1000,\"imageAddress\":null}}"
    delete:
      tags:
      - menu
      summary: 메뉴 삭제
      description: 메뉴 삭제
      operationId: menu-deleteById
      parameters:
      - name: menuId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - menu
      summary: 메뉴 수정
      description: 메뉴 수정
      operationId: menu-update
      parameters:
      - name: menuId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              menu-update:
                value: "{\"name\":\"menu\",\"price\":1000,\"imageAddress\":null}"
      responses:
        "200":
          description: "200"
  /api/v1/notices/{noticeId}:
    get:
      tags:
      - notice
      summary: 공지사항 단건 조회
      description: 공지사항 단건 조회
      operationId: notice-findById
      parameters:
      - name: noticeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                notice-findById:
                  value: "{\"message\":null,\"body\":{\"id\":1,\"title\":\"공지사항 제목\
                    \",\"content\":\"공지사항 내용\",\"createdAt\":\"2023-11-20T14:11:53.025126Z\"\
                    }}"
    delete:
      tags:
      - notice
      summary: 공지사항 삭제
      description: 공지사항 삭제
      operationId: notice-delete
      parameters:
      - name: noticeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - notice
      summary: 공지사항 수정
      description: 공지사항 수정
      operationId: notice-update
      parameters:
      - name: noticeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              notice-update:
                value: "{\"title\":\"newTitle\",\"content\":\"newContent\"}"
      responses:
        "200":
          description: "200"
  /api/v1/stores/{storeId}:
    get:
      tags:
      - store
      summary: 특정 가게 조회
      description: 특정 가게 조회
      operationId: store-findById
      parameters:
      - name: storeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                store-findById:
                  value: "{\"message\":null,\"body\":{\"id\":5,\"displayedName\":\"\
                    명지대학교\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"imageAddress\":null,\"\
                    location\":null,\"categories\":[],\"storeHours\":[]}}"
    delete:
      tags:
      - store
      summary: 특정 가게 정보 삭제
      description: 특정 가게 정보 삭제
      operationId: store-deleteById
      parameters:
      - name: storeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - store
      summary: 특정 가게 정보 업데이트
      description: 특정 가게 정보 업데이트
      operationId: store-updateById
      parameters:
      - name: storeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              store-updateById:
                value: "{\"name\":\"명지대학교\",\"address\":\"서울특별시 서대문구 거북골로 34\",\"\
                  displayName\":\"띵지대\",\"phoneNumber\":{\"phoneNumber\":\"02-300-1656\"\
                  },\"imageAddress\":\"/eatda/public/store/232D8241-C6A9-4AD9-B0EA-56F6DD24BADF.jpg\"\
                  ,\"location\":{\"latitude\":37.5802219,\"longitude\":126.9226047}}"
      responses:
        "204":
          description: "204"
  /api/v1/stores/{storeId}/menu:
    get:
      tags:
      - store
      summary: 가게의 메뉴 전체 조회
      description: 가게의 메뉴 전체 조회
      operationId: store-findAllMenu
      parameters:
      - name: storeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                store-findAllMenu:
                  value: "{\"message\":null,\"body\":[]}"
    post:
      tags:
      - store
      summary: 가게의 메뉴 생성
      description: 가게의 메뉴 생성
      operationId: menu-create
      parameters:
      - name: storeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-categories-category-id486549215'
            examples:
              menu-create:
                value: "{\"name\":\"menu\",\"price\":1000,\"imageAddress\":\"/eatda/public/menu/232D8241-C6A9-4AD9-B0EA-56F6DD24BADF.jpg\"\
                  }"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-categories-category-id486549215'
              examples:
                menu-create:
                  value: "{\"message\":null,\"body\":1}"
components:
  schemas:
    api-v1-categories-category-id486549215:
      type: object